requirements.txt
google-generativeai>=0.7.0
playwright>=1.44.0
icalendar>=5.0.0
python-dateutil>=2.9.0
